Revert "Merge branch 'master' into revive"
This reverts commit 13273f971b16a24707d711717766d4be7e69f10d, reversing
changes made to c27cf7e90c76f4dfa05a47346d0acc8adc4c04f6.
