+++ package.json
@@ -14,9 +14,11 @@
   "author": "LengoLabs",
   "license": "MIT",
   "devDependencies": {
-    "nodemon": "^2.0.15"
+    "nodemon": "^2.0.15",
+    "prisma": "^5.0.0"
   },
   "dependencies": {
+    "@prisma/client": "^5.0.0",
     "@types/express": "^4.17.13",
     "@types/lodash": "^4.14.178",
     "@types/ms": "^0.7.31",
@@ -29,7 +31,6 @@
     "figlet": "^1.5.2",
     "lexure": "^0.17.0",
     "lodash": "^4.17.21",
-    "mongoose": "^6.1.3",
     "ms": "^2.1.3",
     "ts-node": "^10.4.0",
     "typescript": "^4.5.4"
