+++ src/commands/xp/addxp.ts
@@ -18,7 +18,7 @@ import { config } from '../../config';
 import { User, PartialUser, GroupMember } from 'bloxy/dist/structures';
 import { logAction } from '../../handlers/handleLogging';
 import { getLinkedRobloxUser } from '../../handlers/accountLinks';
-import { provider } from '../../database/router';
+import { provider } from '../../database';
 import { findEligibleRole } from '../../handlers/handleXpRankup';
 
 class AddXPCommand extends Command {
