+++ src/commands/xp/removexp.ts
@@ -16,7 +16,7 @@ import { config } from '../../config';
 import { User, PartialUser, GroupMember } from 'bloxy/dist/structures';
 import { logAction } from '../../handlers/handleLogging';
 import { getLinkedRobloxUser } from '../../handlers/accountLinks';
-import { provider } from '../../database/router';
+import { provider } from '../../database';
 
 class RemoveXPCommand extends Command {
     constructor() {
