+++ src/commands/xp/xprankup.ts
@@ -17,7 +17,7 @@ import { config } from '../../config';
 import { User, PartialUser, GroupMember } from 'bloxy/dist/structures';
 import { logAction } from '../../handlers/handleLogging';
 import { getLinkedRobloxUser } from '../../handlers/accountLinks';
-import { provider } from '../../database/router';
+import { provider } from '../../database';
 import { findEligibleRole } from '../../handlers/handleXpRankup';
 
 class XPRankupCommand extends Command {
