+++ src/config.ts
@@ -21,10 +21,6 @@ export const config: BotConfig = {
         actions: '',
         shout: '',
     },
-    database: {
-        enabled: false,
-        type: 'mongodb',
-    },
     api: false,
     maximumRank: 255,
     verificationChecks: false,
