+++ src/events/suspensions.ts
@@ -1,5 +1,5 @@
 import { DatabaseUser } from '../structures/types';
-import { provider } from '../database/router';
+import { provider } from '../database';
 import { robloxGroup } from '../main';
 import { config } from '../config';
