+++ src/structures/DatabaseProvider.ts
@@ -1,6 +1,7 @@
 abstract class DatabaseProvider {
     abstract findUser(query: any);
     abstract findSuspendedUsers();
+    abstract findBannedUsers();
     abstract updateUser(query: any, data: any);
 }
